Fix drawing in dark mode
In dark mode, the canvas is assigned a `height` of `100%`. As a result, it will be assigned the height of the parent element of `100vh`, ignoring the height of the menu. The canvas will be larger than the actual viewport, and therefore the drawing will deviate slightly the further down you move the pointer.